requirements.txt
django
pytest
djangorestframework
psycopg2-binary
